requirements copy 3.txt
_tflow_select==2.2.0
abseil-cpp==20211102.0
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.2
appnope==0.1.3
arrow-cpp==8.0.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
aws-c-common==0.4.57
aws-c-event-stream==0.1.6
aws-checksums==0.1.9
aws-sdk-cpp==1.8.185
backcall==0.2.0
backports==1.0
backports.functools_lru_cache==1.6.4
blas==1.0
blinker==1.5
boost-cpp==1.73.0
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.18.1
ca-certificates==2023.01.10
cachetools==5.3.0
cairo==1.16.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
comm==0.1.2
cryptography==38.0.4
debugpy==1.6.6
decorator==5.1.1
eigen==3.3.7
entrypoints==0.4
executing==1.2.0
expat==2.4.9
ffmpeg==4.2.2
fftw==3.3.9
flask==2.2.2
flatbuffers==23.1.21
flit-core==3.6.0
fontconfig==2.14.1
freetype==2.12.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.2
gast==0.4.0
gettext==0.21.0
gflags==2.2.2
giflib==5.2.1
gitdb==4.0.10
gitpython==3.1.30
glib==2.69.1
glog==0.6.0
gmp==6.2.1
gnutls==3.6.15
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphite2==1.3.14
grpc-cpp==1.46.1
grpcio==1.51.3
gst-plugins-base==1.14.1
gstreamer==1.14.1
gunicorn==20.1.0
h5py==3.8.0
harfbuzz==4.3.0
hdf5==1.10.6
icu==58.2
idna==3.4
importlib-metadata==4.11.3
importlib_resources==5.10.2
intel-openmp==2021.4.0
ipykernel==6.21.1
ipython==8.9.0
ipywidgets==8.0.4
itsdangerous==2.0.1
jedi==0.18.2
jinja2==3.1.2
joblib==1.1.1
jpeg==9e
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.2.0
jupyterlab_widgets==3.0.5
keras==2.11.0
keras-preprocessing==1.1.2
krb5==1.19.3
lame==3.100
lcms2==2.12
lerc==3.0
libboost==1.73.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libclang==15.0.6.1
libcurl==7.87.0
libcxx==14.0.6
libdeflate==1.8
libedit==3.1.20191231
libev==4.33
libevent==2.1.10
libffi==3.4.2
libgfortran==5.0.0
libgfortran5==11.3.0
libiconv==1.16
libidn2==2.3.2
libllvm12==12.0.0
libnghttp2==1.46.0
libopenblas==0.3.21
libopus==1.3.1
libpng==1.6.37
libpq==12.9
libprotobuf==3.20.3
libsodium==1.0.18
libssh2==1.10.0
libtasn1==4.16.0
libthrift==0.15.0
libtiff==4.5.0
libunistring==0.9.10
libuv==1.40.0
libvpx==1.7.0
libwebp==1.2.4
libwebp-base==1.2.4
libxml2==2.9.14
libxslt==1.1.35
lightning-utilities==0.6.0.post0
llvm-openmp==14.0.6
lz4-c==1.9.4
markdown==3.4.1
markdown-it-py==2.1.0
markupsafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.0
mkl==2021.4.0
mkl-service==2.4.0
multidict==6.0.2
nbformat==5.7.3
ncurses==6.4
nest-asyncio==1.5.6
nettle==3.7.3
ninja==1.10.2
ninja-base==1.10.2
nspr==4.33
nss==3.74
numpy==1.23.5
numpy-base==1.23.5
oauthlib==3.2.2
opencv==4.6.0
openh264==2.1.1
openjpeg==2.4.0
openssl==1.1.1t
opt_einsum==3.3.0
orc==1.7.4
packaging==23.0
pandas==1.5.3
parso==0.8.3
pcre==8.45
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
pip==22.3.1
pixman==0.40.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pygments==2.14.0
pyjwt==2.4.0
pympler==1.0.1
pyopenssl==22.0.0
pyrsistent==0.19.3
pysocks==1.7.1
python==3.9.16
python-dateutil==2.8.2
python-fastjsonschema==2.16.2
python-flatbuffers==2.0
python-tzdata==2022.7
python_abi==3.9
pytorch==1.12.1
pytorch-lightning==1.9.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyyaml==6.0
pyzmq==25.0.0
qt-main==5.15.2
qt-webengine==5.15.9
qtwebkit==5.212
re2==2022.04.01
readline==8.2
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
setuptools==65.6.3
six==1.16.0
smmap==3.0.5
snappy==1.1.9
sqlite==3.40.1
stack_data==0.6.2
streamlit==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-base==2.10.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==2.2.0
tk==8.6.12
toml==0.10.2
toolz==0.12.0
torchmetrics==0.11.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.4.0
typing_extensions==4.4.0
tzdata==2022g
tzlocal==4.2
urllib3==1.26.14
utf8proc==2.6.1
validators==0.18.2
watchdog==2.2.1
wcwidth==0.2.6
werkzeug==2.2.2
wheel==0.35.1
widgetsnbextension==4.0.5
wrapt==1.15.0
wtforms==2.3.3
x264==1!157.20191217
xz==5.2.10
yaml==0.2.5
yarl==1.8.1
zeromq==4.3.4
zipp==3.11.0
zlib==1.2.13
zstd==1.5.2
